+++ ICD.Connect.Audio.Biamp/Controls/Volume/IBiampTesiraVolumeDeviceControl.cs
@@ -1,8 +1,0 @@
-﻿using ICD.Connect.Audio.Controls.Volume;
-
-namespace ICD.Connect.Audio.Biamp.Controls.Volume
-{
-	public interface IBiampTesiraVolumeDeviceControl : IBiampTesiraDeviceControl, IVolumeDeviceControl
-	{
-	}
-}